reutilizacion de archivos(NavBar, SubNavbar)
se eliminaron los componentes, NavBar2, NavBar3,NavBar4 el codigo de estos fue agregado a los componentes que habia creado antes agustin(Navbar, SubNavBar)
