+++ src/components/NavBar/index.scss
@@ -1,79 +1,74 @@
-@import '../../scss/modules/variables';
+@import '~bootstrap/scss/functions';
+@import '~bootstrap/scss/variables';
 
-.main-navbar {
-  font-size: 13px;
-  font-family: OpenSans;
-  padding: 8px 16px 0px 16px;
-  display: block;
-  background-color: rgb(243, 243, 243);
-  z-index: 1;
+.image-logo {
+  width: 150px;
 }
 
-.img-logo {
-  width: 200px;
-}
-
-.input-main-search {
-  border-right: none;
-}
-
-.icon-main-search {
-  color: white;
-}
-
-.search-text-group {
+.navbar-toggler {
   border: none;
-  background: $main-color;
 }
 
-.form-control-main-search {
-  border-radius: 10px;
-}
-
-.item,
-.icon-item,
-.icon-link-navbar {
-  color: rgb(102, 102, 102);
-}
-
-.icon-link-navbar:hover {
-  color: rgb(158, 158, 158);
-  text-decoration: none;
-}
-
-.item {
+.responsive-navbar {
   display: flex;
-  margin: 0 1.4em;
-}
-
-.language-dropdown {
-  cursor: pointer;
-}
-
-.language-icon {
-  width: 2em;
-}
-
-.badge-shopping-cart {
-  background-color: $main-color;
-  color: white;
-  position: absolute;
-  top: -40%;
-  left: 100%;
-}
-
-.sub-nav-bar {
-  font-weight: bold;
-  background-color: $main-color;
-  margin: 0 -16px;
-}
-
-.divider-sub-navbar {
-  height: 80%;
-}
-
-.item-nav-bar {
-  display: flex;
-  justify-content: center;
+  list-style: none;
+  border: none;
+  justify-content: space-around;
   align-items: center;
+}
+
+.responsive-navbar .responsive-navbar-item {
+  margin-left: 0.5rem;
+}
+
+@media (min-width: map-get($grid-breakpoints, 'md')) {
+  .responsive-navbar .responsive-navbar-item {
+    margin-left: 2rem;
+    margin-right: 2rem;
+  }
+}
+
+@media (max-width: map-get($grid-breakpoints, 'sm')) {
+  .image-logo {
+    width: 110px;
+  }
+}
+
+.wrapper {
+  max-height: 40;
+  display: flex;
+  overflow-x: auto;
+  background: #ffa200;
+}
+
+.wrapper::-webkit-scrollbar {
+  width: 0;
+  height: 0;
+}
+
+.wrapper .item {
+  min-width: 100px;
+  height: 40px;
+  line-height: 40px;
+  text-align: center;
+  background-color: #ffa200;
+  margin-right: 6px;
+  margin-left: auto;
+}
+.item a {
+  text-decoration: none;
+  color: rgb(255, 255, 255);
+  font-size: 14px;
+  font-family: sans-serif;
+}
+
+.item:hover {
+  background: #e99300;
+}
+
+.divider {
+  background: white;
+  height: 30px;
+  width: 1px;
+  margin-top: 5px;
 }
