--- src/components/Navbar2/index.scss
@@ -1,35 +1,0 @@
-@import '~bootstrap/scss/functions';
-@import '~bootstrap/scss/variables';
-
-.image-logo {
-  width: 150px;
-}
-
-.navbar-toggler {
-  border: none;
-}
-
-.responsive-navbar {
-  display: flex;
-  list-style: none;
-  border: none;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.responsive-navbar .responsive-navbar-item {
-  margin-left: 0.5rem;
-}
-
-@media (min-width: map-get($grid-breakpoints, 'md')) {
-  .responsive-navbar .responsive-navbar-item {
-    margin-left: 2rem;
-    margin-right: 2rem;
-  }
-}
-
-@media (max-width: map-get($grid-breakpoints, 'sm')) {
-  .image-logo {
-    width: 110px;
-  }
-}
