--- src/components/Navbar3/index.scss
@@ -1,38 +1,0 @@
-.wrapper {
-  max-height: 40;
-  display: flex;
-  overflow-x: auto;
-  background: #ffa200;
-}
-
-.wrapper::-webkit-scrollbar {
-  width: 0;
-  height: 0;
-}
-
-.wrapper .item {
-  min-width: 100px;
-  height: 40px;
-  line-height: 40px;
-  text-align: center;
-  background-color: #ffa200;
-  margin-right: 6px;
-  margin-left: auto;
-}
-.item a {
-  text-decoration: none;
-  color: rgb(255, 255, 255);
-  font-size: 14px;
-  font-family: sans-serif;
-}
-
-.item:hover {
-  background: #e99300;
-}
-
-.divider {
-  background: white;
-  height: 30px;
-  width: 1px;
-  margin-top: 5px;
-}
