--- src/components/Navbar4/index.scss
@@ -1,21 +1,0 @@
-.color {
-  background: #00083c;
-}
-
-.letras-color ul li a {
-  color: white;
-  text-decoration: none;
-}
-
-.tam-letra {
-  font-size: 15px;
-}
-.linki:hover {
-  color: #c9c9c9;
-  text-decoration: none;
-}
-
-.linki2:hover {
-  color: #001aff;
-  text-decoration: none;
-}
