+++ src/components/SubNavBar/index.scss
@@ -1,71 +1,21 @@
-.sub-navbar {
-  font-family: OpenSans;
-  display: block;
-  background: rgb(5, 24, 67);
-  font-size: 12px;
+.color {
+  background: #00083c;
 }
 
-.create-store-button {
-  display: inline-block;
+.letras-color ul li a {
   color: white;
-  background: rgb(52, 135, 255);
-  border-radius: 35px;
-  padding: 0.8rem 1rem;
-  font-size: 1rem;
-  font-weight: 400;
-  border: 2px solid rgb(52, 135, 255);
-  transition: color 0.5s, background 0.5s;
-}
-
-.create-store-button:hover {
-  color: rgb(52, 135, 255);
-  background: transparent;
   text-decoration: none;
 }
 
-.create-store-button:active {
-  box-shadow: inset 0 0 5px rgb(52, 135, 255);
-}
-
-.divider-modal {
-  height: 1px;
-  width: 100%;
-  background: #ced4da;
-}
-
-.circle {
-  padding: .5em .9em;
-  border-radius: 50%;
-  border: solid 1px #ced4da;
-}
-
-.circle::before {
-  content: "o";
-
-}
-
-.icon-social-network-modal {
+.tam-letra {
   font-size: 15px;
-  padding: 1em;
-  border-radius: 50%;
-  display: block;
-  box-shadow: 0 3px 5px 0 rgba(0, 0, 0, 0.3);
-  color: white;
+}
+.linki:hover {
+  color: #c9c9c9;
+  text-decoration: none;
 }
 
-.icon-social-network-modal:hover {
-  color: white;
+.linki2:hover {
+  color: #001aff;
+  text-decoration: none;
 }
-
-.icon-social-network-modal.icon-facebook {
-  background-color: rgb(3, 108, 228);
-}
-
-.icon-social-network-modal.icon-twitter {
-  background-color: rgb(3, 169, 244);
-}
-
-.icon-social-network-modal.icon-google {
-  background-color: rgb(255, 182, 53);
-}
-
